fix undefined issue and move tests
